Bug 00000: POC PATCH. I'd like to have DRY RUN and VERBOSE features in shell scripts somewhere

+++ C4/HoldsQueue.pm
@@ -161,9 +161,11 @@ sub GetHoldsQueueItems {
 
 =head2 CreateQueue
 
-  CreateQueue();
+  CreateQueue($params);
 
 Top level function that turns reserves into tmp_holdsqueue and hold_fill_targets.
+In C<$params> hashref can be boolean keys 'verbose' and 'dry_run' for console
+scripts usage.
 
 =cut
 
@@ -171,6 +173,8 @@ sub CreateQueue {
     my $params      = shift;
     my $unallocated = $params->{unallocated};
     my $dbh         = C4::Context->dbh;
+
+    die "DRY_RUN: not capable to dry run yet: needs code refactoring and transactions.\n" if $params->{dry_run};
 
     unless ($unallocated) {
         $dbh->do("DELETE FROM tmp_holdsqueue");    # clear the old table for new info
@@ -195,7 +199,11 @@ sub CreateQueue {
 
     $branches_to_use = load_branches_to_pull_from($use_transport_cost_matrix);
 
+    warn "Branches to use: @$branches_to_use\n" if $params->{verbose} and $branches_to_use;
+
     my $bibs_with_pending_requests = GetBibsWithPendingHoldRequests();
+
+    warn "Bibslios with pending hold requests: @$bibs_with_pending_requests\n" if $params->{verbose} and $bibs_with_pending_requests;
 
     foreach my $biblionumber (@$bibs_with_pending_requests) {
 
@@ -207,6 +215,10 @@ sub CreateQueue {
                 branches_to_use       => $branches_to_use,
                 transport_cost_matrix => $transport_cost_matrix,
                 unallocated           => $unallocated
+            },
+            {
+                dry_run => $params->{dry_run},
+                verbose => $params->{verbose},
             }
         );
 
@@ -1178,7 +1190,9 @@ are allowed to be passed to avoid calculating them many times inside loops.
 =cut
 
 sub update_queue_for_biblio {
-    my ($args) = @_;
+    my $args = shift;
+    my $params = shift;
+
     my $biblio_id = $args->{biblio_id};
     my $result;
 
@@ -1216,6 +1230,8 @@ sub update_queue_for_biblio {
         CreatePicklistFromItemMap($item_map);
         AddToHoldTargetMap($item_map);
     }
+
+    warn "Created and added hold to target map for items: " . join(', ', keys %$item_map) . "\n" if $params->{verbose} and %$item_map;
 
     return $result;
 }
